Fix bug in the ray tracer
Basically, the ray tracer assumed that the planes given to the function
planes_lines() all intersected the ray in the positive direction.
This assumption was made because planes intersecting the ray in the
negative direction should have been pruned via the rtree. However, if
the ray's origin is too close to any triangle, the triangle won't be
pruned even if intersects the ray in the negative direction.

In order to fix this, this commit adds a new function, planes_rays(),
that performs the check explicitly. The ray tracers now call this
function instead of planes_lines() by default.