Bugfix for multi-mesh file loads using assimp
The main load method assumed that all loaders returned a single
kwargs dictionary, but the assimp loaders return a list of kwargs
dictionaries (as many filetypes return several types of meshes).
This causes the assimp loaders to fail.

This patch fixes this problem by explicitly checking if the return
value from the loader is a dict or a list. We shouldn't change the
assimp behavior, as some file formats (like Wavefront .obj) can
certainly store multiple meshes in the same file.