Add URDF exporter
This takes a directory as an argument rather than a file, and it
uses the convex decomposition module to generate a collision-checkable
approximation of the mesh and exports a .urdf file, all the nearly-convex
sub-meshes, and a Gazebo model.config to the target directory.

Eventually, I'd like to figure out a clean way to add this to the
standard export_mesh function, but since it operates a bit differently
from the standard exporters (i.e. exports many files into a directory),
I've left it as-is for now.