Fix bug when updating vertices with a mask
Basically, the given code in update_vertices() is designed to update the
vertex list by filtering it out with a mask. In addition, it attempts
to filter the vertex normals by the same mask. However, the given
code will never accomplish this -- the newly-masked vertex normals
won't be the same shape as the old vertex list, so setting
vertex_normals will result in a no-op, leaving the vertex normals
unchanged. Assigning to the vertices first, then to the vertex normals
fixes this bug.