Support differing list-lengths in ASCII-PLYs
PLY allows defining list-properties, where every instance of the
property can have a different length. This can for instance be useful
when storing submeshes in the metadata of a PLY-file.

trimesh only supported PLY-files with list-properties, if every
instance of the list had the same length. Properties with lists of
differing lengths let the parser crash at the `vstack`-call in
`ply.ply_ascii`, as arrays of differing size cannot be stacked.

The implementation at hand adds support for lists of differing length.

The drawback of the new implementation is an increased load-time for
elements containing lists. I tried to mitigate this effect by checking
whether an element has no list or just one where every instance has the
same length and using the numpy-based algorithm to get the data in this
case, while iterating all rows, if there's more than one list or one
list with instances of differing lengths.