reworking to_planar logic
Fixing minor issues with code but also reworking logic of how to_2D and normal inputs are handled. Updated util.unitize function implemented to make it easier to handle major axis detection and get rid of floating point errors.

Currently a ValueError is raised when points are planar but not aligned with major axis. Could possibly in the future instead detect if they're planar and proceed without a major axis alignment.