Speed up proximity.signed_distance
proximity.signed_distance was using an expensive raycast which was comparing all candidate points to the mesh to determine the sign of the distance. For a large number of cases the raycast is not necessary as we already have a filtered list of triangles that we can compare the points to first.

The approach implemented here first checks if the projection of each point lies within the closest triangle, if so, the point, projection, and normal are used to determine the sign of the distance. Where the projection lies outside the triangle, the previous method continues to be used.

For my use case where a large number of points project cleanly in to the mesh triangles, I observed this technique to be an order of magnitude faster.

Tests were updated/added.