Fix export of ASCII ply with attributes
**Summary**
Exporting a mesh with attributes with ASCII encoding would result in a malformed ply file which specified the attributes in the header but did not include them in the data section. This fix was facilitated by a new `structured_array_to_string` method. I originally looked at generalizing the existing method but decided upon a bespoke method instead due to the implications of mixed data type rows.

Also fixed a couple bugs in the existing `array_to_string` method

**Test Plan**
- Added tests for `array_to_string`
- Added tests for `structured_array_to_string`