calculate integral mean curvature
This function calculates the integral mean curvature of a mesh. With this addition, all four Minkowski functionals for a surface mesh embedded in 3d can be calculated by Trimesh. (Volume, Area, Integral Mean Curvature, Euler number)

In the most general case, the lengths of the two numpy arrays,
self.face_adjacency_angles
and
self.edges_unique_length,

may not be equal. This could be checked with an 'assert.'