Use available triangulation engines in tests
In test_path_sweep and test_triangulate_plumbing, use all available triangulation engines (and no unavailable ones). This specifically allows the tests to run when the 'earcut' engine is available but the non-free 'triangle' engine is not. It also provides coverage of the 'earcut' engine when 'triangle' (the implicit default) *is* available.