Allow splitting OBJ by 'o' and 'g' tags
Previously it was only possible to group objects by material. Instead, it is now possible to
split objects by different group names or object names. This is similar to blender, which also
allows splitting OBJ by group and object. The implementation was straightforward, as previously
loading 'o' and 'g' tags was already implemented, but nothing seemed to use the functionality.
The one thing I had to add was group name in face_tuples, and update the functionality of
group_by_material to also handle objects and groups.